Added notch filters

+++ run_pipeline.py
@@ -12,16 +12,17 @@ import shutil
 from glob import glob
 from pathlib import Path
 import json
+import matplotlib.pyplot as plt
 
 from ibllib.ephys import ephysqc
 from ibllib.ephys.spikes import ks2_to_alf, sync_spike_sorting
 from ibllib.pipes.ephys_tasks import (EphysCompressNP1, EphysSyncPulses, EphysSyncRegisterRaw,
                                       EphysPulses)
 from brainbox.metrics.single_units import spike_sorting_metrics
+from phylib.io.model import get_template_params
+from phylib.utils._misc import write_python
 
-import spikeinterface.extractors as se
-import spikeinterface.preprocessing as spre
-from spikeinterface.sorters import run_sorter, get_default_sorter_params
+import spikeinterface.full as si
 
 from powerpixel_utils import load_neural_data
 
@@ -47,7 +48,7 @@ if isfile(join(dirname(realpath(__file__)), 'spikesorter_param_files',
                    f'{settings_dict["SPIKE_SORTER"]}_params.json'), 'r') as openfile:
         sorter_params = json.load(openfile)
 else:
-    sorter_params = get_default_sorter_params(settings_dict['SPIKE_SORTER'])
+    sorter_params = si.get_default_sorter_params(settings_dict['SPIKE_SORTER'])
 
 # Initialize Matlab engine for bombcell package
 if settings_dict['RUN_BOMBCELL']:
@@ -94,6 +95,7 @@ for root, directory, files in os.walk(settings_dict['DATA_FOLDER']):
                 
         # Loop over multiple probes 
         probes = glob(join(root, 'raw_ephys_data', 'probe*'))
+        probe_done = np.zeros(len(probes)).astype(bool)
         for i, probe_path in enumerate(probes):
             print(f'\nStarting preprocessing of {split(probe_path)[-1]}')
             
@@ -101,6 +103,7 @@ for root, directory, files in os.walk(settings_dict['DATA_FOLDER']):
             this_probe = split(probe_path)[1]
             if isdir(join(root, this_probe + id_str)):
                 print('Probe already processed, moving on')
+                probe_done[i] = True
                 continue
             
             # Create probe sync file
@@ -117,37 +120,44 @@ for root, directory, files in os.walk(settings_dict['DATA_FOLDER']):
             # Load in recording
             if len(glob(join(probe_path, '*.cbin'))) > 0:
                 # Recording is already compressed by a previous run, loading in compressed data
-                rec = se.read_cbin_ibl(probe_path)
+                rec = si.read_cbin_ibl(probe_path)
             else:
-                rec = se.read_spikeglx(probe_path, stream_id=f'imec{split(probe_path)[-1][-1]}.ap')
+                rec = si.read_spikeglx(probe_path, stream_id=f'imec{split(probe_path)[-1][-1]}.ap')
             
             # Apply high-pass filter
             print('\nApplying high-pass filter.. ')
-            rec_filtered = spre.highpass_filter(rec)
-            
+            rec_filtered = si.highpass_filter(rec, ftype='bessel', dtype='float32')
+                        
             # Correct for inter-sample phase shift
             print('Correcting for phase shift.. ')
-            rec_shifted = spre.phase_shift(rec_filtered)
+            rec_shifted = si.phase_shift(rec_filtered)
             
             # Do common average referencing
             print('Performing common average referencing.. ')
-            rec_comref = spre.common_reference(rec_filtered)
+            rec_comref = si.common_reference(rec_filtered)
             
-            # Detect and interpolate over bad channels
             print('Detecting and interpolating over bad channels.. ')
-            bad_channel_ids, all_channels = spre.detect_bad_channels(rec_comref)
             
-            # If there are too many bad channels, skip the interpolation step
-            prec_bad_ch = np.sum(all_channels == 'noise') / all_channels.shape[0]
-            if prec_bad_ch < (1/3):
-                rec_interpolated = spre.interpolate_bad_channels(rec_shifted, bad_channel_ids)
-                print(f'{np.sum(all_channels == "noise")} ({prec_bad_ch*100:.0f}%) bad channels')
-            else:
-                rec_interpolated = rec_comref
-                print(f'{np.sum(all_channels == "noise")} ({prec_bad_ch*100:.0f}%) bad channels,',
-                      'skipping the interpolation step')
+            # Detect dead channels
+            bad_channel_ids, all_channels = si.detect_bad_channels(rec_filtered, seed=42)
+            prec_dead_ch = np.sum(all_channels == 'dead') / all_channels.shape[0]
+            print(f'{np.sum(all_channels == "dead")} ({prec_dead_ch*100:.0f}%) dead channels')
+            dead_channel_ids = rec_filtered.get_channel_ids()[all_channels == 'dead']
+            prec_out_ch = np.sum(all_channels == 'out') / all_channels.shape[0]
+            print(f'{np.sum(all_channels == "out")} ({prec_out_ch*100:.0f}%) channels outside of the brain')
+            out_channel_ids = rec_filtered.get_channel_ids()[all_channels == 'out']
             
-            """
+            # Detect noisy channels
+            bad_channel_ids, all_channels = si.detect_bad_channels(rec_comref, method='mad',
+                                                                   std_mad_threshold=3, seed=42)
+            prec_noise_ch = np.sum(all_channels == 'noise') / all_channels.shape[0]
+            print(f'{np.sum(all_channels == "noise")} ({prec_noise_ch*100:.0f}%) noise channels')
+            noisy_channel_ids = rec_comref.get_channel_ids()[all_channels == 'noise']
+            
+            # Interpolate over bad channels          
+            rec_interpolated = si.interpolate_bad_channels(rec_shifted, np.concatenate((
+                dead_channel_ids, noisy_channel_ids, out_channel_ids)))
+                        
             # If there are multiple shanks, do destriping per shank
             print('Destriping.. ')
             if np.unique(rec_interpolated.get_property('group')).shape[0] > 1:
@@ -156,30 +166,57 @@ for root, directory, files in os.walk(settings_dict['DATA_FOLDER']):
                 rec_split = rec_interpolated.split_by(property='group')
                 rec_destriped = []
                 for sh in range(len(rec_split)):
-                    rec_destriped.append(spre.highpass_spatial_filter(rec_split[sh]))
+                    rec_destriped.append(si.highpass_spatial_filter(rec_split[sh]))
                 
                 # Merge back together
-                rec_final = si.aggregate_channels(rec_destriped)
+                rec_destriped = si.aggregate_channels(
+                    rec_destriped, renamed_channel_ids=rec_interpolated.get_channel_ids())
                 
             else:
                 
                 # Do destriping for the whole probe at once
-                rec_final = spre.highpass_spatial_filter(rec_interpolated)
-            """
+                rec_destriped = si.highpass_spatial_filter(rec_interpolated)
             
-            # If there are multiple shanks, skip destriping for now
-            print('Destriping.. ')
-            if np.unique(rec_interpolated.get_property('group')).shape[0] > 1:
-                rec_final = rec_interpolated
+            # Plot spectral density
+            print('Calculating power spectral density')
+            data_chunk = si.get_random_data_chunks(rec_destriped, num_chunks_per_segment=1,
+                                                   chunk_size=30000, seed=42)
+            fig, ax = plt.subplots(figsize=(10, 7))
+            for tr in data_chunk.T:
+                p, f = ax.psd(tr, Fs=rec_destriped.sampling_frequency, color="b")
+            plt.savefig(join(probe_path, 'power spectral density.jpg'), dpi=600)
+            
+            # Apply notch filter 
+            if isfile(join(probe_path, 'notch_filter.json')):
+                
+                # Load in notch filter settings
+                with open(join(probe_path, 'notch_filter.json'), 'r') as openfile:
+                    notch_filter = json.load(openfile)
+                    
+                # Apply filters
+                rec_notch = rec_destriped
+                for freq, q in zip(notch_filter['FREQ'], notch_filter['Q']):
+                    print(f'Applying notch filter at {freq} Hz..')
+                    rec_notch = si.notch_filter(rec_notch, freq=freq, q=q)
+                    
+                # Plot spectral density
+                print('Calculating power spectral density')
+                data_chunk = si.get_random_data_chunks(rec_notch, num_chunks_per_segment=1,
+                                                       chunk_size=30000, seed=42)
+                fig, ax = plt.subplots(figsize=(10, 7))
+                for tr in data_chunk.T:
+                    p, f = ax.psd(tr, Fs=rec_destriped.sampling_frequency, color="b")
+                plt.savefig(join(probe_path, 'power spectral density after notch filter.jpg'), dpi=600)
+                
+                rec_final = rec_notch
+                
             else:
-                # Do destriping for the whole probe at once
-                print('Destriping.. ')
-                rec_final = spre.highpass_spatial_filter(rec_interpolated)
-                        
+                rec_final = rec_destriped
+                
             # Run spike sorting
             try:
                 print(f'\nStarting {split(probe_path)[-1]} spike sorting at {datetime.now().strftime("%H:%M")}')
-                sort = run_sorter(
+                sort = si.run_sorter(
                     settings_dict['SPIKE_SORTER'],
                     rec_final,
                     output_folder=join(probe_path, settings_dict['SPIKE_SORTER'] + id_str),
@@ -193,6 +230,9 @@ for root, directory, files in os.walk(settings_dict['DATA_FOLDER']):
                 logf = open(os.path.join(probe_path, 'error_log.txt'), 'w')
                 logf.write(str(err))
                 logf.close()
+                
+                # Delete empty sorting directory
+                shutil.rmtree(join(probe_path, settings_dict['SPIKE_SORTER'] + id_str))
                 
                 # Continue with next recording
                 continue
@@ -224,9 +264,17 @@ for root, directory, files in os.walk(settings_dict['DATA_FOLDER']):
                                  probe_path,
                                  nargout=0)
             
-            # Export spike sorting to alf files
-            if not isdir(join(root, this_probe)):
-                os.mkdir(join(root, this_probe))
+            # Change the pointer in phy to the raw data from the temporary recording.dat
+            # to the original bin file
+            phy_params = get_template_params(sorter_out_path / 'params.py')
+            phy_params['dat_path'] = str(bin_path / ap_file)
+            phy_params['dir_path'] = str(phy_params['dir_path'])
+            phy_params['dtype'] = str(phy_params['dtype'])
+            write_python(sorter_out_path / 'params.py', phy_params)
+            
+            # Export as alf
+            if not isdir(alf_path):
+                os.mkdir(alf_path)
             ks2_to_alf(sorter_out_path, bin_path, alf_path)
             
             # Move LFP power etc. to the alf folder
@@ -274,6 +322,7 @@ for root, directory, files in os.walk(settings_dict['DATA_FOLDER']):
             if isfile(join(sorter_out_path, 'recording.dat')):
                 os.remove(join(sorter_out_path, 'recording.dat'))
                 
+            """
             # Compress raw data
             if len(glob(join(root, 'raw_ephys_data', this_probe, '*ap.cbin'))) == 0:
                 print('Compressing raw binary file')
@@ -288,8 +337,11 @@ for root, directory, files in os.walk(settings_dict['DATA_FOLDER']):
                 except:
                     print('Could not remove uncompressed ap bin file, delete manually')
                     continue
+            """
             
+            probe_done[i] = True
             print(f'Done! At {datetime.now().strftime("%H:%M")}')
         
-        # Delete process_me.flag
-        os.remove(os.path.join(root, 'process_me.flag'))
+        # Delete process_me.flag if all probes are processed
+        if np.sum(probe_done) == len(probes):
+            os.remove(os.path.join(root, 'process_me.flag'))
